update : code splitted author's hero section
